requirements.txt
asgiref==3.8.1
Django==5.1.6
psycopg==3.2.4
psycopg-binary==3.2.4
sqlparse==0.5.3
typing_extensions==4.12.2
crispy-bootstrap5==2024.10
django-allauth==65.4.1
environs[django]==14.1.1
pandas==2.2.3
django-tables2==2.7.5
django-filter==25.1
django-autocomplete-light==3.12.1
django-widget-tweaks==1.5.0

